move helper scripts to scripts dir
