+++ README
@@ -545,13 +545,13 @@ It should print ``Hello, world!``.
 
 FUZZING
 -------
-The repository includes a simple Swift libFuzzer harness in ``fuzz.swift``.
-The helper script ``scripts/swift_fuzz.py`` compiles the harness and forwards
+The repository includes a simple Swift libFuzzer harness in ``scripts/fuzz/fuzz.swift``.
+The helper script ``scripts/fuzz/swift_fuzz.py`` compiles the harness and forwards
 any arguments to the resulting fuzzer executable. A quick smoke test can be
 run as follows:
 
 ```
-$ python3 scripts/swift_fuzz.py -runs=1
+$ python3 scripts/fuzz/swift_fuzz.py -runs=1
 ```
 
 Additional libFuzzer flags may be provided after the script name.
@@ -593,7 +593,7 @@ the numeric values::
 
 Missing or malformed files fall back to built-in defaults. Writing a
 new value to one of these files immediately updates the active
-constant. The ``flow_pid.py`` helpers ``set_kp()``, ``set_ki()`` and
+constant. The ``scripts/flow_pid.py`` helpers ``set_kp()``, ``set_ki()`` and
 ``set_kd()`` provide a convenient API for adjusting the parameters.
 
 For testing the base directory can be overridden with the environment
@@ -604,14 +604,14 @@ with a Python example is provided in
 STREAMS LOGGING
 ---------------
 Python STREAMS modules log messages using ``strlog_json()`` from
-``streams_log.py``.  The helper outputs a JSON object containing the
+``scripts/streams_log.py``.  The helper outputs a JSON object containing the
 timestamp, log level and message.  Extra keyword arguments become
 fields in the record.  ``strlog()`` continues to print plain text for
 older modules.
 
 Example::
 
-    from streams_log import strlog_json
+    from scripts.streams_log import strlog_json
     strlog_json("info", "module loaded", name="xor_encrypt")
 
 AUTOMATED SETUP WITH CODEX
